LayoutTests/http/tests/cache/main-resource-304-reload-expected.txt

Verify that document loaded with 304 response can be reloaded correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentDocument.body.textContent is "body text"
PASS iframe.contentDocument.body.textContent is "body text"
Reloading
PASS iframe.contentDocument.body.textContent is "body text"
PASS successfullyParsed is true

TEST COMPLETE
